+++ dev/inference/dalle_mini
