--- dev/notebooks/README.md
@@ -1,5 +1,0 @@
-# Notebooks
-
-These notebooks were used during development.
-
-TODO: This section requires some refactor and clean up.
